% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross validation.R
\name{CV_criterion}
\alias{CV_criterion}
\title{Computes the cross-validation criterion for given model dimensions m,
tilde_m, and H. Note that the splitting of the data is random and happens in this
function. This consecutive calls of this function will yield different
results.
Input: The input variables are all identical to laguerre_cross_validation,
see there for details. For this function only the model dimensions
need to specified in addition.
Output: The value of the cross-validation criterion.}
\usage{
CV_criterion(
  m,
  m_tilde,
  H,
  X,
  X_s,
  type,
  Y,
  Delta,
  tau,
  starting_beta,
  trials,
  nfolds,
  verbose = FALSE,
  link = "exp"
)
}
\description{
Computes the cross-validation criterion for given model dimensions m,
tilde_m, and H. Note that the splitting of the data is random and happens in this
function. This consecutive calls of this function will yield different
results.
Input: The input variables are all identical to laguerre_cross_validation,
see there for details. For this function only the model dimensions
need to specified in addition.
Output: The value of the cross-validation criterion.
}
