% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{graph}
\alias{graph}
\title{A function that plots the estimated quantile and sigma.
The arguments are:
H: The estimated coefficients of the Hermite series for sigma.
beta: The estimated regression coefficients.
X: The design matrix of the regression.
Y: The vector of responses.
Delta: The censorship indicator.}
\usage{
graph(fit)
}
\description{
A function that plots the estimated quantile and sigma.
The arguments are:
H: The estimated coefficients of the Hermite series for sigma.
beta: The estimated regression coefficients.
X: The design matrix of the regression.
Y: The vector of responses.
Delta: The censorship indicator.
}
