% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross validation.R
\name{laguerre_cross_validation_het}
\alias{laguerre_cross_validation_het}
\title{A function that finds the best model dimensions by cross-validation. It does not perform an exhaustive search, rather, it starts
with the simplest model, and moves up in complexity by increasing th dimension that gives the greatest decrease in the cross-validation
criterion. It iterates until a local optimum is found.}
\usage{
laguerre_cross_validation_het(
  X,
  X_s,
  type,
  Y,
  Delta,
  tau,
  starting_beta = FALSE,
  trials = 32,
  nfolds = 5,
  verbose = FALSE,
  link = "exp",
  h_step
)
}
\description{
A function that finds the best model dimensions by cross-validation. It does not perform an exhaustive search, rather, it starts
with the simplest model, and moves up in complexity by increasing th dimension that gives the greatest decrease in the cross-validation
criterion. It iterates until a local optimum is found.
}
