% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{laguerre_estimator_het}
\alias{laguerre_estimator_het}
\title{Perform Laguerre Censored Quantile Regression}
\usage{
laguerre_estimator_het(
  m,
  m_tilde,
  H,
  X,
  X_s = 0,
  type,
  Y,
  Delta,
  tau,
  starting_beta = FALSE,
  trials = 32,
  verbose = 1,
  link = "exp"
)
}
\arguments{
\item{m}{the order of the series expansion on the positive real axis.}

\item{m_tilde}{the order of the series expansion on the negative real axis.}

\item{H}{the order of the series expansion for sigma (modulo link function).}

\item{X}{the design matrix. The first column must be a vector of 1's (for the intercept).}

\item{X_s}{the matrix of variables that sigma depends upon. Do not include an intercept.}

\item{type}{A vector of character values indicating the type of polynomials to be used in the expansion of sigma. Current options are "Hermite", "Laguerre", and "Legendre". If multiple variables appear in X_s, then 'type' must have as many elements as variables are in X_s. The tensor product basis will be used in this multivariate case.}

\item{Y}{The vector of observed values (Y = min(T,C)).}

\item{Delta}{The vector of censorship status (1 for observed, 0 for censored).}

\item{tau}{The quantile level to estimate.}

\item{starting_beta}{The initial values of the coefficients. Defaults to FALSE.}

\item{trials}{The number of initial values to try in the optimization of the log-likelihood function. Defaults to 32.}

\item{verbose}{Controls for output during optimization. If 0, no output is printed. If 1, a progress bar is shown indicating progress. If 2, a message is shown if an error occured during a trial. If 3, both progress bar and error alert are shown.}

\item{link}{The link function in sigma = link(orthogonal expansion). Options are "exp" for an exponential link, "quad" for a quadratic link, or a custom link function. A custom argument must be a list of two functions. The first entry is the link function, and the second entry is its derivative.}
}
\description{
A function that estimates the model with heteroskedasticity.
}
\details{
The parameters m and m_tilde represent the order of the Laguerre series used for the positive and
negative parts of the error density, as described in the paper by Alexander Kreiss and Ingrid van Keilegom.
The parameter H represents the order of the series expansion for sigma, the heteroskedasticity function of the
regression model. Current options for this orthogonal expansion are Hermite, Laguerre and Legendre polynomials.
}
